+++ src/rogallo/widgets/__init__.py
@@ -2,20 +2,16 @@
 
 ##############################################################################
 # Local imports.
-from .bookmarks import BookmarksViewer
-from .client_certificates import ClientCertificateManager
 from .command_line import CommandLine
-from .history import HistoryViewer
+from .sidepanel import SidePanel
 from .toolbar import Toolbar
 from .viewer import Viewer
 
 ##############################################################################
 # Exports.
 __all__ = [
-    "BookmarksViewer",
-    "ClientCertificateManager",
     "CommandLine",
-    "HistoryViewer",
+    "SidePanel",
     "Toolbar",
     "Viewer",
 ]